+++ test/analyticsSessionEvent.spec.js
@@ -1,7 +1,6 @@
 const { expect } = require('chai');
 const apiVideo = require('../lib');
 const { ITEMS_TOTAL } = require('./api');
-const { version } = require('../package.json');
 
 describe('AnalyticsSessionEvent ressource', () => {
   describe('get without parameters', () => {
@@ -9,18 +8,6 @@ describe('AnalyticsSessionEvent ressource', () => {
 
     it('Does not throw', async () => {
       await client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x');
-    });
-
-    it('Sends good request', () => {
-      client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x').catch(() => {});
-      expect(client.analyticsSessionEvent.browser.lastRequest).to.deep.equal({
-        url: 'https://ws.api.video/analytics/sessions/psx1x1x1x1x1x1x1x1x1x/events?pageSize=100&currentPage=1',
-        method: 'GET',
-        headers: {
-          'User-Agent': `api.video SDK (nodejs; v:${version})`,
-        },
-        json: true,
-      });
     });
   });
 
@@ -35,18 +22,6 @@ describe('AnalyticsSessionEvent ressource', () => {
       await client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x', parameters);
     });
 
-    it('Sends good request', () => {
-      client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x', parameters).catch(() => {});
-      expect(client.analyticsSessionEvent.browser.lastRequest).to.deep.equal({
-        url: 'https://ws.api.video/analytics/sessions/psx1x1x1x1x1x1x1x1x1x/events?currentPage=1&pageSize=25',
-        method: 'GET',
-        headers: {
-          'User-Agent': `api.video SDK (nodejs; v:${version})`,
-        },
-        json: true,
-      });
-    });
-
     it('Retrieves only the first page', async () => {
       const analyticSessionEvent = await client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x', parameters);
       expect(analyticSessionEvent.events).to.have.lengthOf(parameters.pageSize);
@@ -59,18 +34,6 @@ describe('AnalyticsSessionEvent ressource', () => {
     it('Does not throw', async () => {
       await client.analyticsSessionEvent
         .get('psx1x1x1x1x1x1x1x1x1x');
-    });
-
-    it('Sends good request', () => {
-      client.analyticsSessionEvent.get('psx1x1x1x1x1x1x1x1x1x').catch(() => {});
-      expect(client.analyticsSessionEvent.browser.lastRequest).to.deep.equal({
-        url: 'https://ws.api.video/analytics/sessions/psx1x1x1x1x1x1x1x1x1x/events?pageSize=100&currentPage=1',
-        method: 'GET',
-        headers: {
-          'User-Agent': `api.video SDK (nodejs; v:${version})`,
-        },
-        json: true,
-      });
     });
 
     it('Retrieves all pages', async () => {