+++ test/tokens.spec.js
@@ -1,25 +1,11 @@
 const { expect } = require('chai');
 const apiVideo = require('../lib');
-const { version } = require('../package.json');
 
 describe('Tokens ressource', () => {
   describe('generate', () => {
     const client = new apiVideo.Client({ apiKey: 'test' });
     it('Does not throw', async () => {
       await client.tokens.generate();
-    });
-
-    it('Sends good request', () => {
-      client.tokens.generate().catch(() => {});
-      expect(client.tokens.browser.lastRequest).to.deep.equal({
-        url: 'https://ws.api.video/tokens',
-        method: 'POST',
-        headers: {
-          'User-Agent': `api.video SDK (nodejs; v:${version})`,
-        },
-        body: {},
-        json: true,
-      });
     });
   });